Add test standard loop
