+++ src/ProcessMaker/Nayra/Contracts/Bpmn/StandardLoopCharacteristicsInterface.php
@@ -8,7 +8,6 @@ namespace ProcessMaker\Nayra\Contracts\Bpmn;
 interface StandardLoopCharacteristicsInterface extends LoopCharacteristicsInterface
 {
     const BPMN_PROPERTY_TEST_BEFORE = 'testBefore';
-    const BPMN_PROPERTY_TEST_AFTER = 'testAfter';
     const BPMN_PROPERTY_LOOP_MAXIMUM = 'loopMaximum';
     const BPMN_PROPERTY_LOOP_CONDITION = 'loopCondition';
     
@@ -29,35 +28,19 @@ interface StandardLoopCharacteristicsInterface extends LoopCharacteristicsInterf
     public function setTestBefore(bool $testBefore);
 
     /**
-     * Gets as testAfter
-     *
-     * @return boolean
-     */
-    public function getTestAfter();
-
-    /**
-     * Sets a new testAfter
-     *
-     * @param bool $testAfter
-     *
-     * @return self
-     */
-    public function setTestAfter(bool $testAfter);
-
-    /**
      * Gets as loopMaximum
      *
-     * @return FormalExpressionInterface
+     * @return string
      */
     public function getLoopMaximum();
 
     /**
      * Sets a new loopMaximum
      *
-     * @param FormalExpressionInterface $loopMaximum
+     * @param string $loopMaximum
      * @return self
      */
-    public function setLoopMaximum(FormalExpressionInterface $loopMaximum);
+    public function setLoopMaximum(string $loopMaximum);
 
     /**
      * Gets as loopCondition
