+++ tests/Feature/Patterns/files/Loop_ConditionOnly.bpmn
@@ -2,9 +2,9 @@
 <bpmn2:definitions xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                    xmlns:bpmn2="http://www.omg.org/spec/BPMN/20100524/MODEL"
                    xsi:schemaLocation="http://www.omg.org/spec/BPMN/20100524/MODEL http://www.omg.org/spec/BPMN/20100501/BPMN20.xsd"
-                   id="Loop_TaskTest"
-                   targetNamespace="http://dsg.wiai.uniba.de/betsy/bpmn/loop_Task">
-    <bpmn2:process id="Loop_Task" isExecutable="true">
+                   id="Loop_ConditionOnlyTest"
+                   targetNamespace="http://dsg.wiai.uniba.de/betsy/bpmn/loop_ConditionOnly">
+    <bpmn2:process id="Loop_ConditionOnly" isExecutable="true">
         <bpmn2:startEvent id="StartEvent_1" name="Start">
             <bpmn2:outgoing>SequenceFlow_2</bpmn2:outgoing>
         </bpmn2:startEvent>
