+++ tests/Feature/Patterns/files/Loop_NoIteration_TestBeforeFalse.json
@@ -1,0 +1,12 @@
+[
+    {
+        "comment": "The StandardLoopCharacteristics class defines looping behavior based on a boolean condition. The Activity will loop as long as the boolean condition is true.",
+        "startEvent": "StartEvent",
+        "data" : {},
+        "result": [
+            "start",
+            "LoopTask",
+            "end"
+        ]
+    }
+]
