+++ tests/Feature/Patterns/files/Loop_NoIteration_TestBeforeTrue.bpmn
@@ -2,13 +2,15 @@
 <bpmn2:definitions xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
                    xmlns:bpmn2="http://www.omg.org/spec/BPMN/20100524/MODEL"
                    xsi:schemaLocation="http://www.omg.org/spec/BPMN/20100524/MODEL http://www.omg.org/spec/BPMN/20100501/BPMN20.xsd"
-                   id="Loop_TaskTest"
-                   targetNamespace="http://dsg.wiai.uniba.de/betsy/bpmn/loop_Task">
-    <bpmn2:process id="Loop_Task" isExecutable="true">
-        <bpmn2:startEvent id="StartEvent_1" name="Start">
+                   id="Loop_NoIteration_TestBeforeTrueTest"
+                   targetNamespace="http://dsg.wiai.uniba.de/betsy/bpmn/loop_NoIteration_TestBeforeTrue">
+    <bpmn2:process id="Loop_NoIteration_TestBeforeTrue" isExecutable="true">
+
+        <bpmn2:startEvent id="StartEvent" name="Start">
             <bpmn2:outgoing>SequenceFlow_2</bpmn2:outgoing>
         </bpmn2:startEvent>
-        <bpmn2:sequenceFlow id="SequenceFlow_2" name="" sourceRef="StartEvent_1" targetRef="start"/>
+
+        <bpmn2:sequenceFlow id="SequenceFlow_2" name="" sourceRef="StartEvent" targetRef="start"/>
 
         <bpmn2:scriptTask id="start" name="Start">
             <bpmn2:incoming>SequenceFlow_2</bpmn2:incoming>
@@ -16,20 +18,20 @@
             <bpmn2:script>CREATE_LOG_FILE</bpmn2:script>
         </bpmn2:scriptTask>
 
-        <bpmn2:sequenceFlow id="SequenceFlow_1" name="" sourceRef="start" targetRef="ScriptTask_1"/>
+        <bpmn2:sequenceFlow id="SequenceFlow_1" name="" sourceRef="start" targetRef="LoopTask"/>
 
-        <bpmn2:scriptTask id="ScriptTask_1" name="task1" scriptFormat="application/x-betsy">
+        <bpmn2:scriptTask id="LoopTask" name="Loop">
             <bpmn2:incoming>SequenceFlow_1</bpmn2:incoming>
             <bpmn2:outgoing>SequenceFlow_5</bpmn2:outgoing>
-            <bpmn2:standardLoopCharacteristics>
-                <bpmn2:loopCondition xsi:type="bpmn2:tFormalExpression">test->data['integerVariable']&lt;3</bpmn2:loopCondition>
+            <bpmn2:standardLoopCharacteristics testBefore="true">
+                <bpmn2:loopCondition xsi:type="bpmn2:tFormalExpression">false</bpmn2:loopCondition>
             </bpmn2:standardLoopCharacteristics>
-            <bpmn2:script>return ['integerVariable' => $data['integerVariable']+1];</bpmn2:script>
+            <bpmn2:script>SCRIPT_task1</bpmn2:script>
         </bpmn2:scriptTask>
 
-        <bpmn2:sequenceFlow id="SequenceFlow_5" name="" sourceRef="ScriptTask_1" targetRef="end"/>
+        <bpmn2:sequenceFlow id="SequenceFlow_5" name="" sourceRef="LoopTask" targetRef="end"/>
 
-        <bpmn2:scriptTask id="end" name="task2">
+        <bpmn2:scriptTask id="end" name="End">
             <bpmn2:incoming>SequenceFlow_5</bpmn2:incoming>
             <bpmn2:outgoing>SequenceFlow_4</bpmn2:outgoing>
             <bpmn2:script>SCRIPT_task2</bpmn2:script>
@@ -40,5 +42,6 @@
         <bpmn2:endEvent id="EndEvent_2">
             <bpmn2:incoming>SequenceFlow_4</bpmn2:incoming>
         </bpmn2:endEvent>
+
     </bpmn2:process>
 </bpmn2:definitions>
