+++ tests/Feature/Patterns/files/Loop_Task.json
@@ -1,6 +1,6 @@
 [
     {
-        "comment": "The StandardLoopCharacteristics class defines looping behavior based on a boolean condition. The Activity will loop as long as the boolean condition is true.",
+        "comment": "1. The StandardLoopCharacteristics class defines looping behavior based on a boolean condition. The Activity will loop as long as the boolean condition is true.",
         "startEvent": "StartEvent_1",
         "data": {
             "integerVariable": 4
@@ -12,7 +12,7 @@
         ]
     },
     {
-        "comment": "The StandardLoopCharacteristics class defines looping behavior based on a boolean condition. The Activity will loop as long as the boolean condition is true.",
+        "comment": "2. The StandardLoopCharacteristics class defines looping behavior based on a boolean condition. The Activity will loop as long as the boolean condition is true.",
         "startEvent": "StartEvent_1",
         "data": {
             "integerVariable": 3
@@ -24,7 +24,7 @@
         ]
     },
     {
-        "comment": "The StandardLoopCharacteristics class defines looping behavior based on a boolean condition. The Activity will loop as long as the boolean condition is true.",
+        "comment": "3. The StandardLoopCharacteristics class defines looping behavior based on a boolean condition. The Activity will loop as long as the boolean condition is true.",
         "startEvent": "StartEvent_1",
         "data": {
             "integerVariable": 2
